requirements.txt
pylint
pytest
pytest-cov
flask
gunicorn
requests
pandas
psycopg2-binary